ai-service/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
pydantic==2.10.0
python-multipart==0.0.9
python-dotenv==1.0.1
httpx==0.27.0
aiofiles==24.1.0
loguru==0.7.2
pandas==2.2.2
scikit-learn==1.5.1
joblib==1.4.2
psutil==6.0.0
